Added setPrices Action
