+++ package.json
@@ -32,6 +32,7 @@
     "ethers": "6.13.2",
     "graphql": "^16.9.0",
     "hardhat": "^2.22.9",
-    "rollup": "^4.9.1"
+    "rollup": "^4.9.1",
+    "rollup-plugin-visualizer": "^5.12.0"
   }
 }
